pyproject.toml
[build-system]
requires = ["setuptools>=65.0.0", "wheel>=0.37.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kobo2pandas"
version = "0.9.0"
description = "Desde la API de Kobo a pandas.DataFrame"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Diego Mora", email = "morabdiego@gmail.com"}
]
maintainers = [
    {name = "Diego Mora", email = "morabdiego@gmail.com"}
]
keywords = [
    "kobo",
    "kobotoolbox",
    "data-extraction",
    "pandas",
    "excel",
    "humanitarian",
    "surveys",
    "data-processing"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.5.0",
    "openpyxl>=3.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.28.0"
]

[project.optional-dependencies]
dev = [
    "build>=0.8.0",
    "twine>=4.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/morabdiego/kobo2pandas"
Documentation = "https://github.com/morabdiego/kobo2pandas#readme"
Repository = "https://github.com/morabdiego/kobo2pandas"
"Bug Tracker" = "https://github.com/morabdiego/kobo2pandas/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["kobo2pandas*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
kobo2pandas = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=kobo2pandas --cov-report=term-missing --cov-report=html"
